requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
dj-database-url==2.3.0
Django==5.1.6
django-allauth==65.5.0
django-ckeditor==6.7.2
django-environ==0.12.0
django-js-asset==3.0.1
django-taggit==6.1.0
gunicorn==23.0.0
idna==3.10
packaging==24.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.3.0
whitenoise==6.9.0
Pillow==10.2.0